backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
firebase-admin==6.3.0
livekit==0.11.1
livekit-agents==0.7.2
openai==1.6.1
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2